requirements.txt
fastapi
uvicorn
python-dotenv
SQLAlchemy
passlib[bcrypt]
python-jose
pydantic
